requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
ansible==10.7.0
ansible-core==2.17.13
anyio==3.7.1
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==24.3.0
aws==0.2.5
awscli==1.44.5
babel==2.17.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.4
boto==2.49.0
botocore==1.42.15
Brlapi==0.8.3
captum==0.8.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
contourpy==1.3.0
cryptography==45.0.5
cupshelpers==1.0
cycler==0.12.1
Cython==3.0.2
dataclasses-json==0.6.7
dbus-python==1.2.18
debugpy==1.8.16
decorator==5.2.1
dedoc==2.3.2
dedoc-utils==0.3.8
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docutils==0.19
duplicity==0.8.21
elastic-transport==8.17.1
elasticsearch==8.19.0
email-validator==2.3.0
exceptiongroup==1.2.2
executing==2.2.1
fabric==3.2.2
fasteners==0.14.1
fastjsonschema==2.21.2
filelock==3.18.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.7.0
future==0.18.2
greenlet==3.2.3
grpcio==1.74.0
h11==0.14.0
h2o==3.46.0.5
hf-xet==1.1.5
holidays==0.80
httpcore==1.0.7
httplib2==0.20.2
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
imageio==2.37.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
imutils==0.5.4
invoke==2.2.1
ipykernel==6.30.1
ipython==8.37.0
ipython_genutils==0.2.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.1
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==0.2.1
kazam==1.4.5
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-elasticsearch==0.3.2
langchain-ollama==0.2.2
langchain-text-splitters==0.3.8
langgraph==0.2.66
langgraph-checkpoint==2.1.1
langgraph-sdk==0.1.74
langsmith==0.3.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libevdev==0.5
lightning-utilities==0.15.2
limits==5.4.0
lockfile==0.12.2
louis==3.20.0
lxml==5.4.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.1.4
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==2.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
neuralprophet==0.9.0
notebook==7.4.5
notebook_shim==0.2.4
ntlm-auth==1.4.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
olefile==0.46
ollama==0.4.6
opencv-python==4.5.5.64
orjson==3.10.15
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.5
passlib==1.7.4
pdf.tocgen==1.3.4
pdf2image==1.10.0
pdfminer.six==20231228
pexpect==4.8.0
piexif==1.1.3
pillow==11.2.1
platformdirs==4.4.0
plotly==6.3.0
prettytable==3.17.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
prophet==1.1.7
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==20.0.0
pyasn1==0.6.1
PyAudio==0.2.11
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic-settings==2.7.1
pydantic_core==2.14.1
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pylogbeat==2.0.1
pylzma==0.5.0
pymacaroons==0.13.0
PyMuPDF==1.25.5
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==4.1.0
PyPDF2==1.27.0
pyRFC3339==1.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==0.8.11
python-dotenv==1.0.1
python-json-logger==3.3.0
python-Levenshtein==0.12.2
python-logstash-async==2.7.0
python-magic==0.4.27
pytorch-lightning==2.5.5
pytz==2025.2
pyudev==0.22.0
PyWavelets==1.8.0
pywinrm==0.3.0
pyxdg==0.27
PyYAML==6.0.2
pyzmq==27.0.2
rarfile==4.0
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==1.0.0
resolvelib==0.8.1
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==3.3
rpds-py==0.24.0
rsa==4.7.2
s3transfer==0.16.0
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.2
SecretStorage==3.3.1
Send2Trash==1.8.3
simsimd==6.5.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
ssh-import-id==5.11
stack-data==0.6.3
stanio==0.5.1
starlette==0.27.0
sympy==1.14.0
systemd-python==234
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
torch==2.8.0
torchmetrics==1.8.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.8.0
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
uvloop==0.22.1
wadllib==1.3.6
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.14
wrapt==1.17.2
xdg==5
xgbfir==0.3.1
xgboost==1.7.6
xkit==0.0.0
xlrd==1.2.0
xlsxwriter==3.2.5
xmltodict==0.12.0
yarl==1.18.3
zipp==1.0.0
zstandard==0.23.0
